move annotations to their own module
