+++ apollo-annotations/gradle.properties
@@ -1,0 +1,3 @@
+POM_ARTIFACT_ID=apollo-annotations
+POM_NAME=Apollo GraphQL Annotations
+POM_DESCRIPTION=Apollo GraphQL Annotations
