+++ tests/models-operation-based/src/jvmTest/kotlin/test/TestBuildersTest.kt
@@ -6,7 +6,7 @@ import codegen.models.MergedFieldWithSameShapeQuery
 import codegen.models.test.AllPlanetsQuery_TestBuilder.Data
 import codegen.models.test.HeroAndFriendsWithTypenameQuery_TestBuilder.Data
 import codegen.models.test.MergedFieldWithSameShapeQuery_TestBuilder.Data
-import com.apollographql.apollo3.api.ApolloExperimental
+import com.apollographql.apollo3.annotations.ApolloExperimental
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
